Merged PR 19: Ably Proof of Concept (POC)
Implemented actions and reducers to update the state based on events published by the service through an Ably channel.
Implemented a `PubSubService` to manage subscriptions to Ably channels.
Updated the `TripResultsScreen` to start a search, display data received on the channel, and stop a search when the user presses on "Back".

Related work items: #84